+++ docs/walkthroughs/01-installing-slate.md
@@ -80,7 +80,7 @@ const App = () => {
 
 Next up is to render a `<Slate>` context provider.
 
-The provider component keeps track of your Slate editor, its plugins, its value, its selection, and any changes that occur. It **must** be rendered above any `<Editable>` components. But it can also provide the editor state to other components like toolbars, menus, etc. using the `useSlate` hook.
+The provider component keeps track of your Slate editor, its plugins, its value, its selection, and any changes that occur. It **must** be rendered above any `<Editable>` components. But it can also provide the editor state to other components like toolbars, menus, etc. using the `useEditor` hook.
 
 ```jsx
 const App = () => {
