+++ packages/slate-react/src/hooks/use-editor.tsx
@@ -1,20 +1,27 @@
-import { useContext } from 'react'
+import { createContext, useContext } from 'react'
+import { Editor } from 'slate'
+import { ReactEditor } from '../plugin/react-editor'
 
-import { EditorContext } from './use-editor-static'
+/**
+ * A React context for sharing the editor object, in a way that re-renders the
+ * context whenever changes occur.
+ */
+
+export const SlateContext = createContext<[ReactEditor] | null>(null)
 
 /**
  * Get the current editor object from the React context.
- * @deprecated Use useEditorStatic instead.
  */
 
-export const useEditor = () => {
-  const editor = useContext(EditorContext)
+export const useEditor = (): Editor => {
+  const context = useContext(SlateContext)
 
-  if (!editor) {
+  if (!context) {
     throw new Error(
-      `The \`useEditor\` hook must be used inside the <Slate> component's context.`
+      `The \`useEditor\` hook must be used inside the <SlateProvider> component's context.`
     )
   }
 
+  const [editor] = context
   return editor
 }
