+++ packages/slate-react/src/index.ts
@@ -9,12 +9,12 @@ export { DefaultLeaf } from './components/leaf'
 export { Slate } from './components/slate'
 
 // Hooks
-export { useEditor } from './hooks/use-editor'
+export { _useEditor } from './hooks/_use-editor'
 export { useEditorStatic } from './hooks/use-editor-static'
 export { useFocused } from './hooks/use-focused'
 export { useReadOnly } from './hooks/use-read-only'
 export { useSelected } from './hooks/use-selected'
-export { useSlate } from './hooks/use-slate'
+export { useEditor } from './hooks/use-editor'
 
 // Plugin
 export { ReactEditor } from './plugin/react-editor'
