+++ site/examples/richtext.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useMemo, useState } from 'react'
 import isHotkey from 'is-hotkey'
-import { Editable, withReact, useSlate, Slate } from 'slate-react'
+import { Editable, withReact, useEditor, Slate } from 'slate-react'
 import {
   Editor,
   Transforms,
@@ -146,7 +146,7 @@ const Leaf = ({ attributes, children, leaf }) => {
 }
 
 const BlockButton = ({ format, icon }) => {
-  const editor = useSlate()
+  const editor = useEditor()
   return (
     <Button
       active={isBlockActive(editor, format)}
@@ -161,7 +161,7 @@ const BlockButton = ({ format, icon }) => {
 }
 
 const MarkButton = ({ format, icon }) => {
-  const editor = useSlate()
+  const editor = useEditor()
   return (
     <Button
       active={isMarkActive(editor, format)}
